Split Aff v2 route helpers
